Force download of benchmark mapping file
With this commit we always download the mapping file. In the past
it happened that we had to update the mapping file due to changes
in Elasticsearch but Rally users had to manually remove the file
in order to get a working benchmark environment again.

By always downloading the mapping file we ensure that it is
current. While this is not an ideal solution, users are at least
not blocked when a mapping change has been made.