Allow to benchmark a binary distribution
With this commit we add a new pipeline 'from-distribution' and a
new command line parameter 'distribution-version' which allow
users to benchmark an official binary distribution of
Elasticsearch.

This is intended for end users which are more likely to be
interested in performance characteristics of releases in
contrast to core developers which also want to know performance
characteristics of development versions.

Closes #63