Allow specifying multiple mappings
With this commit we remove the constraint that only one index and
one type is allowed. We can now specify multiple indices and in
each index multiple types. For each type there is a mandatory
mapping file and an optional data file.

Existing tracks have been minimally adapted and the 'tiny' track
serves as an example of the new track specification format.

Closes #73