Gather GC stats for whole benchmarking run
With this commit we capture GC stats not only for the index part but for the
whole benchmark run. This brings the GC stats in line with total times reported
by GC logs and other tools.

We also gather stats now via a telemetry device freeing up the benchmark driver
from this burden.

Closes #65