Honor HTTP proxy settings on connection probing
Previously, when Rally probed for an Internet connection, it just opened
a plain socket connection. This is problematic when a user connects via
a proxy to the Internet because Rally will then conclude that there is
no working Internet connection at all.

With this commit, we respect the HTTP proxy setting on connection
probing too.

Closes #245