Allow non-deterministic operation schedules
With this commit we introduce a new `schedule` property that can be used
to specify an arbitrary operation schedule. Out of the box, Rally ships
with the `deterministic` and `poisson` schedules (default:
`deterministic`). Users are also enabled to implement their own
schedules and we also add documentation on how to do that.

Closes #251