Keep cancellation state if racecontrol has detected it
With this commit we immediately cancel the benchmark if cancellation is
detected in racecontrol (coordinator) instead of first forwarding the
request to the driver and relying that it will respond correctly. With
this change, Rally behaves more straightforward during cancellation even
in situations where there is a problem with the driver.

Relates #298