Distribute load-test driver
With this commit we distribute the load test driver. Clients will be
assigned in a round-robin strategy to available load driver hosts.

With this commit we also use only relative paths in tracks for a
machine-agnostic representation. On each load generator we amend the
(local) track (copy) and use correct absolute paths.

Closes #257