Correct index time calculation for multiple laps
With this commit we correctly determine the indexing times related
statistics even when running for multiple laps. Previously we determined
the values at start and end and subtracted them. The problem with this
approach is that right after we got the value at the start, Rally
recreated the test indices thus resetting all indexing times stats to
zero. Then we subtracted the end value from the high value from the
start and constantly got a negative value.

Now we skip this subtraction and warn the user instead about a
inconsistent cluster state at the beginning of the benchmark if these
values are not zero.