Remove callable formatter
The logging configuration will be written to a separate file (for use
with the Thespian Director) whose contents can only be a dictionary,
not generic runnable Python code, so a callable formatter cannot
easily be defined.  The commented-out form *could* be used if that
import path object was guaranteed to exist for all target
environments, but that's not a current assumption.

The functionality loss here is that times are now reported as local
times instead of UTC times.  Another workaround for this would be to
have the rally startup Thespian only after setting the TZ environment
variable to UTC.  Alternatively, the %z timezone specification could
be included in the datefmt.