Use ActorTypeDispatcher to simplify message dispatch.
The ActorTypeDispatcher automatically implements the "if
isinstance(msg, X):" functionality and captures that in an entrypoint
of "def receiveMsg_X(self, msg, sender):".  This helps simplify the
incoming message dispatching, including removing syntax nesting levels
from the code.

The mechanic shows the minimal conversion, whereas the driver and
racecontrol show a more complete conversion.

The try/catch wrapped around the previous receiveMessage handling is
replaced by Thespian's automatic sending of a PoisonMessage if the
actor encounters an exception when processing a message.  The caller
then needs to have a "receiveMsg_PoisonMessage" entrypoint providing
similar functionality to the "receiveMsg_BenchmarkFailed" entrypoint.