Store structured user tags for results and race
Previously we have stored user tags in a structured manner only for raw
metrics (in the `meta` block) but we have only stored the raw string in
the results and race indices in the metrics store.

With this commit we introduce a new property `user-tags` for both which
contain all provided user tags in a structured manner.

Closes #392