Be even more backwards compatible with bash function parameters
94e43130f06492a4d27ca4aee566342f0856ca64 fixed incompatibilities with
older bash versions such as the one found in macOS, however, it's not
bullet proof and still fails e.g. with Ubuntu 16.04's
bash (`4.3.48(1)).

It seems that declaring functions[1] like
`install_esrally_with_setuptools ()` fixes this issue, but for now we
opt for the solution we had earlier using `${1}` for the parameter
inside the function. Additional tests may need to be added in the
future.

[1]
https://www.gnu.org/software/bash/manual/bashref.html#Shell-Functions