Disable automatic preparation of offline distro
With this commit we disable the automatic invocation of the script that
creates the Rally offline distribution. We disable it because the new
pypi API caches quite heavily and the Rally package will not show up
immediately after the release. Therefore, creating this distribution is
an additional manual step in the release process for the time being.