Gather cluster-level metrics in driver
With this commit we gather cluster-level metrics in the driver instead of the
mechanic. As these metrics are gathered via API calls there is no need to gather
them on the very same machine were an Elasticsearch node is running. Instead, it
defines clearer boundaries in between these two components.

Relates #697
Relates #779