Store Rally revision in metrics metadata (#931)
Rally already stores the version string as metrics metadata. However,
for unreleased versions we intermingle the version string with the git
revision making it hard to determine the revision automatically.

With this commit we store the git revision of Rally separately from the
version for race and results metadata. This allows users to derive
Rally's git revision more easily.