Allow to abort on fatal errors (#997)
With this commit we add a third option to the command line parameter
`on-error`. So far it has been possible to either continue or abort on
error. Now we also allow to continue except for errors that indicate
that the cluster is unreachable (e.g. because it died with an OOME).