Allow to downsample request metrics (#1001)
With this commit we introduce a new setting
`metrics.request.downsample.factor` that allows to control that only
every N-th request metric sample is recorded in the metrics store (this
does not affect throughput calculation which will still consider every
sample). This allows to reduce the number of samples which can be useful
when benchmarks are run with many hundreds or thousands of clients which
would produce a very high number of samples.