Fetch artifact revision lazily on build (#1028)
With this commit we avoid touching the source repository when a specific
git revision is provided for an artifact (with `--revision`) and the
artifact is already present in the source artifact cache. This helps in
situations where an external process has put artifacts into Rally's
source artifact cache from a fork (in which case the artifact revision
will not be present in the main repo and the resolution would fail).