Ensure that sleep tasks are timed (#1121)
With this commit we ensure that also the `sleep` task is included in
time measurements. Regular runners in Rally issue requests against
Elasticsearch and time measurement is done via callbacks in the request
lifecycle but the `sleep` runner is different because it only waits for
a certain time without any calls.