Log trace of uncaught exception in load generator (#1162)
With this commit we enhance the handlers for uncaught exception handlers
in the load generator to log the full exception trace before forwarding
the exception *message* to the coordinating process. Previously, the
coordinating process would then log the exception message but we lost
the actual trace which makes diagnosis harder.