Document meta-data returned by operations (#1169)
With this commit we add a new `meta data` section to each operation in
the track reference documentation which documents which meta data are
returned. This information can be used to add assertions for custom
tracks. We also enhance a couple of existing runners so that they return
a `dict` structure instead of a tuple because only `dicts` are supported
for assertions.

Relates #1157