Support gradual ramp-up (#1266)
With this commit we allow users to ramp-up load gradually by specifying
the task property `ramp-up-time-period`. If a non-zero value is
specified, Rally will gradually add clients during that time period
until the target client count as specified by `clients` is reached. This
reduces the potential for initial load spikes when running with many
concurrent clients.

Closes #1195