Retry incomplete HTTP downloads (#1521)
Currently when we did not download enough data, we just abort the
benchmark. Instead, let's ask urllib3 to retry it per our retry policy.

This does not change the existing Rally logic around Content-Length as
it's possible to specify the expected length and I don't want this fix
to change other behavior.