Pass plugin-params also for modules (#1640)
In #1624 we've ensured that plugins that have turned into modules are
now skipped from the mandatory plugin check. However, we've also not
passed any plugin parameters to the provisioner and this made post
installation hooks fail, if they expected parameters to be passed on the
command line via `--plugin-params` (e.g. `repository-gcs`). With this
commit we make sure that `--plugin-params` is still honored even in
these cases.

Relates #1624