Add looped mode to bulk operation (#1830)
Adds looped mode to bulk operation. That is useful in benchmarks where
we want Rally to ingest for a specific amount of time even if corpus is
not big enough to sustain indexing at a specific rate for this time.

This mode is meant to be used together with `time-period` or
`iterations` specified at the task level. Otherwise Rally will never
complete the task.