Preserve pre-existing `rally-metrics` index templates by default (#1912)
It handles issue #1900: Rally should not overwrite pre existing templates by default

When opening an EsMetricsStore, it creates the index template if any of following is true:

- the index template doesn't exist
- `reporting/datastore.overwrite_existing_templates` option is true and there are differences
  between existing and requested template.

It will preserve existing template on all the other cases.
It logs a warning when an existing index template is being replaced.
It logs index template differences between the existing one (if any) and configured one.